Merge Head changes into my fork

+++ tests/trajectory-planner/circular-arcs/test-all.sh
@@ -1,9 +1,12 @@
 #!/bin/bash
 set -o monitor
+if [ ! -e "$1" ]
+then
+    echo "INI file not specified, aborting auto test!"
+    exit 1
+fi
 cp position.blank position.txt
-touch constraints.log
-rm constraints.log
 #Assume build without TP debug logging enabled
 linuxcnc $1 > test.log &
-python run_all_tests.py
+python2 run_all_tests.py
 fg
